refactor(webserver): separate handlers
